requirements.txt
gradio
cryptography
uuid
logging
python-dotenv
groq
gtts
soundfile
numpy